requirements.txt
pandas==2.0.3
notebook
fastapi==0.110.0
uvicorn==0.28.0
psycopg2-binary==2.9.9